+++ src/UnwritableCategorySourceException.php
@@ -9,8 +9,12 @@ use Throwable;
 
 final class UnwritableCategorySourceException extends Exception
 {
-    public function __construct(string $categoryName, ?Throwable $previous = null)
+    public function __construct(string $categoryName, int $code = 0, ?Throwable $previous = null)
     {
-        parent::__construct(sprintf('Unable to write into category with name "%s".', $categoryName), 0, $previous);
+        parent::__construct(
+            sprintf('The category source "%s" does not support writing.', $categoryName),
+            $code,
+            $previous
+        );
     }
 }